Move CopyrightholderAdmin, OwnerAdmin and PublisherAdmin
