+++ journal/wagtail_hooks.py
@@ -2,23 +2,23 @@ from django.http import HttpResponseRedirect
 from django.urls import path
 from django.utils.translation import gettext as _
 from wagtail import hooks
-from wagtail_modeladmin.options import (
-    ModelAdmin,
-    ModelAdminGroup,
-    modeladmin_register,
+from wagtail.snippets.models import register_snippet
+from wagtail.snippets.views.snippets import (
+    CreateView,
+    SnippetViewSet,
+    SnippetViewSetGroup,
 )
+from wagtail_modeladmin.options import ModelAdmin
 from wagtail_modeladmin.views import CreateView, EditView
-from wagtail.admin.panels import FieldPanel, InlinePanel
+
+from config.menu import get_menu_order
+
 from . import models
 from .button_helper import IndexedAtHelper
 from .views import import_file, validate
-from config.menu import get_menu_order
 
 COLLECTION_TEAM = "Collection Team"
 JOURNAL_TEAM = "Journal Team"
-
-from wagtail.snippets.models import register_snippet
-from wagtail.snippets.views.snippets import SnippetViewSet, CreateView, SnippetViewSetGroup
 
 
 class OfficialJournalCreateViewSnippet(CreateView):
@@ -320,90 +320,6 @@ class ArticleSubmissionFormatCheckListAdmin(ModelAdmin):
     menu_order = get_menu_order("article_subm")
 
 # modeladmin_register(ArticleSubmissionFormatCheckListAdmin)
-
-
-class OwnerAdmin(ModelAdmin):
-    model = models.Owner
-    menu_icon = "folder"
-    menu_order = 300
-    menu_label = _("Owner")
-    add_to_settings_menu = False  # or True to add your model to the Settings sub-menu
-    exclude_from_explorer = (
-        False  # or True to exclude pages of this type from Wagtail's explorer view
-    )
-    list_display = ("institution",)
-    search_fields = (
-        "institution__institution_identification__name",
-        "institution__institution_identification__acronym",
-        "institution__level_1",
-        "institution__level_2",
-        "institution__level_3",
-    )
-    list_export = (
-        "institution__institution_identification__name",
-        "institution__institution_identification__acronym",
-        "institution__level_1",
-        "institution__level_2",
-        "institution__level_3",
-        "location",
-    )
-    export_filename = "owner"
-
-
-class CopyrightholderAdmin(ModelAdmin):
-    model = models.CopyrightHolder
-    menu_icon = "folder"
-    menu_order = 400
-    menu_label = _("Copyrightholder")
-    add_to_settings_menu = False  # or True to add your model to the Settings sub-menu
-    exclude_from_explorer = (
-        False  # or True to exclude pages of this type from Wagtail's explorer view
-    )
-    list_display = ("institution",)
-    search_fields = (
-        "institution__institution_identification__name",
-        "institution__institution_identification__acronym",
-        "institution__level_1",
-        "institution__level_2",
-        "institution__level_3",
-    )
-    list_export = (
-        "institution__institution_identification__name",
-        "institution__institution_identification__acronym",
-        "institution__level_1",
-        "institution__level_2",
-        "institution__level_3",
-        "location",
-    )
-    export_filename = "copyrightholder"
-
-
-class PublisherAdmin(ModelAdmin):
-    model = models.Publisher
-    menu_icon = "folder"
-    menu_order = 500
-    menu_label = _("Publisher")
-    add_to_settings_menu = False  # or True to add your model to the Settings sub-menu
-    exclude_from_explorer = (
-        False  # or True to exclude pages of this type from Wagtail's explorer view
-    )
-    list_display = ("institution",)
-    search_fields = (
-        "institution__institution_identification__name",
-        "institution__institution_identification__acronym",
-        "institution__level_1",
-        "institution__level_2",
-        "institution__level_3",
-    )
-    list_export = (
-        "institution__institution_identification__name",
-        "institution__institution_identification__acronym",
-        "institution__level_1",
-        "institution__level_2",
-        "institution__level_3",
-        "location",
-    )
-    export_filename = "Publisher"
 
 
 @hooks.register("register_admin_urls")
